+++ crates/wasmtime/src/runtime/vm/gc/enabled/trace_infos.rs
@@ -6,63 +6,8 @@
 
 use crate::hash_map::{Entry, HashMap};
 use crate::module::RegisteredModuleId;
-use crate::vm::{GcStoreTraceState, TraceInfo};
-use core::hash::BuildHasher;
+use crate::vm::{GcStoreTraceState, NopHasher, TraceInfo};
 use wasmtime_environ::{ModuleInternedTypeIndex, VMSharedTypeIndex};
-
-/// A hasher that doesn't hash, for use in the trace-info hash map, where we are
-/// just using scalar keys and aren't overly concerned with collision-based DoS.
-#[derive(Default)]
-struct NopHasher(u64);
-
-impl BuildHasher for NopHasher {
-    type Hasher = Self;
-
-    #[inline]
-    fn build_hasher(&self) -> Self::Hasher {
-        NopHasher::default()
-    }
-}
-
-impl core::hash::Hasher for NopHasher {
-    #[inline]
-    fn finish(&self) -> u64 {
-        self.0
-    }
-
-    #[inline]
-    fn write(&mut self, bytes: &[u8]) {
-        let mut hash = self.0.to_ne_bytes();
-        let n = hash.len().min(bytes.len());
-        hash[..n].copy_from_slice(bytes);
-        self.0 = u64::from_ne_bytes(hash);
-    }
-
-    #[inline]
-    fn write_u8(&mut self, i: u8) {
-        self.write_u64(i.into());
-    }
-
-    #[inline]
-    fn write_u16(&mut self, i: u16) {
-        self.write_u64(i.into())
-    }
-
-    #[inline]
-    fn write_u32(&mut self, i: u32) {
-        self.write_u64(i.into())
-    }
-
-    #[inline]
-    fn write_u64(&mut self, i: u64) {
-        self.0 = i;
-    }
-
-    #[inline]
-    fn write_usize(&mut self, i: usize) {
-        self.write_u64(i.try_into().unwrap());
-    }
-}
 
 #[derive(Clone, Copy)]
 enum TraceInfoLoc {
